adding some files for better project  - travis  - eslint  - mocha
