+++ .eslintrc
@@ -1,0 +1,6 @@
+{
+  "env": {
+    "mocha": true
+  },
+  "extends": "kapgel/legacy"
+}
